requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
openai
tiktoken
numpy
